# Improved performance of primitive comparisons.