Implementation based on [this comment](https://github.com/apache/arrow/pull/9759#discussion_r598154125) from @yordan-pavlov.

Performance:

```
eq Float32              time:   [70.990 us 71.775 us 72.636 us]
                        change: [-12.665% -10.799% -8.9867%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) high mild
  1 (1.00%) high severe

eq scalar Float32       time:   [26.832 us 27.174 us 27.540 us]
                        change: [-52.665% -51.655% -50.525%] (p = 0.00 < 0.05)
                        Performance has improved.
```